backend/requirements.txt
fastapi
tortoise-orm
uvicorn
aiosqlite  # Required for async SQLite support